Check default values with `confirmation` [See #137]
Addresses part of
[#137](https://github.com/DockYard/ember-changeset-validations/issues/137)

The `confirmation` validation only looks to the `changes` object
for the state of other fields. However, this may result in a
false positive when the initial value of the field matches
what's in the `confirmation` field, because `changes` only contains
fields with new data that is valid.

This doesn't fix the issue of an invalid field that matches
what's in the confirmation field. Technically, if an invalid field
matches what's in the confirmation field there should be no error.
A more fundamental change inside of `ember-changeset` is required
for this to work properly.