before/after/onOrBefore/onOrAfter accept function (#281)
The Date validator's before/after/onOrBefore/onOrAfter options now
accept a function that returns a date, which is re-evaluated everytime
the validator runs. This allows the date validation logic to
change in between validations, which is useful for situations where
date validation logic is based on date values that may change in between
validations.